rename player and enemy
